move and rename hasOverloadSignatures
